less is more
